Extract memoryStore
